# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Devora AI Code Generator API
  description: |
    **Devora** is an AI-powered code generation SaaS platform that transforms natural language descriptions into production-ready web applications.

    ## Features
    - Multi-agent AI orchestration for specialized code generation
    - Full-stack Next.js 14+ application generation
    - Real-time streaming progress via Server-Sent Events
    - GitHub export and Vercel deployment integration
    - Stripe payment processing
    - Persistent memory with Memori SDK

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Base URL
    - Development: `http://localhost:8000`
    - Production: `https://api.devora.ai`

    ## Rate Limiting
    - **Free Plan**: 10 requests/minute
    - **Pro Plan**: 100 requests/minute
    - **Enterprise**: Unlimited

  version: 3.0.0
  contact:
    name: Devora Support
    email: support@devora.ai
    url: https://devora.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.devora.ai
    description: Production server
...
paths:
...
  /api/projects:
    get:
      tags:
        - Projects
      summary: List all projects
      description: Get all projects for the authenticated user
      operationId: listProjects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Projects
      summary: Create a new project
      description: Create a new project with files and metadata
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '200':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/projects/{project_id}:
    get:
      tags:
        - Projects
      summary: Get a project by ID
      description: Retrieve a specific project with all its files
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Projects
      summary: Update a project
      description: Update project metadata, files, or conversation history
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Projects
      summary: Delete a project
      description: Permanently delete a project and all its files
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Project deleted successfully"
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and return JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique project identifier
  schemas:
...
    ProjectFile:
      type: object
      required:
        - name
        - content
        - language
      properties:
        name:
          type: string
          example: "app/page.tsx"
        content:
          type: string
          example: "export default function Page() { return <div>Hello</div> }"
        language:
          type: string
          enum: [typescript, javascript, html, css, markdown, json]
    ConversationMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    Project:
      type: object
      required:
        - id
        - name
        - files
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "My SaaS App"
        description:
          type: string
          nullable: true
        files:
          type: array
          items:
            $ref: '#/components/schemas/ProjectFile'
        conversation_history:
          type: array
          items:
            $ref: '#/components/schemas/ConversationMessage'
        project_type:
          type: string
          enum: [saas, ecommerce, blog, dashboard, api, landing]
          nullable: true
        github_repo_url:
          type: string
          nullable: true
          example: "https://github.com/user/repo"
        vercel_url:
          type: string
          nullable: true
          example: "https://my-app.vercel.app"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/ProjectFile'
          default: []
        project_type:
          type: string
          enum: [saas, ecommerce, blog, dashboard, api, landing]
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            plan:
              type: string
              enum: [free, pro, enterprise]
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Error message"
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid authentication credentials"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Project not found"
...